Extract rest api view model to independent project
