+++ src/Microsoft.DocAsCode.DataContracts.RestApi/Microsoft.DocAsCode.Build.RestApi.xproj
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="__ToolsVersion__" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="..\Shared\xproj.base.props" />
+  <PropertyGroup Label="Globals">
+    <ProjectGuid>C3F43329-9344-4E6A-8698-6C50861AE150</ProjectGuid>
+  </PropertyGroup>
+  <PropertyGroup>
+    <SchemaVersion>2.0</SchemaVersion>
+  </PropertyGroup>
+  <Import Project="$(VSToolsPath)\DNX\Microsoft.DNX.targets" Condition="'$(VSToolsPath)' != ''" />
+</Project>
