+++ src/Microsoft.DocAsCode.DataContracts.RestApi/RestApiItemViewModelBase.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.DocAsCode.Build.RestApi.ViewModels
+namespace Microsoft.DocAsCode.DataContracts.RestApi
 {
     using System;
     using System.Collections.Generic;
@@ -14,33 +14,38 @@ namespace Microsoft.DocAsCode.Build.RestApi.ViewModels
     using Microsoft.DocAsCode.YamlSerialization;
 
     [Serializable]
-    public class RestApiTagViewModel : IOverwriteDocumentViewModel
+    public class RestApiItemViewModelBase : IOverwriteDocumentViewModel
     {
         [YamlMember(Alias = Constants.PropertyName.Uid)]
         [JsonProperty(Constants.PropertyName.Uid)]
         [MergeOption(MergeOption.MergeKey)]
         public string Uid { get; set; }
 
-        [YamlMember(Alias = Constants.PropertyName.Conceptual)]
-        [JsonProperty(Constants.PropertyName.Conceptual)]
-        public string Conceptual { get; set; }
-
-        [YamlMember(Alias = Constants.PropertyName.Documentation)]
-        [JsonProperty(Constants.PropertyName.Documentation)]
-        public SourceDetail Documentation { get; set; }
+        [YamlMember(Alias = "htmlId")]
+        [JsonProperty("htmlId")]
+        [MergeOption(MergeOption.Ignore)]
+        public string HtmlId { get; set; }
 
         [YamlMember(Alias = "name")]
         [JsonProperty("name")]
         public string Name { get; set; }
 
+        [YamlMember(Alias = Constants.PropertyName.Conceptual)]
+        [JsonProperty(Constants.PropertyName.Conceptual)]
+        public string Conceptual { get; set; }
+
         [YamlMember(Alias = "description")]
         [JsonProperty("description")]
         public string Description { get; set; }
 
-        [YamlMember(Alias = "htmlId")]
-        [JsonProperty("htmlId")]
+        [YamlMember(Alias = "summary")]
+        [JsonProperty("summary")]
+        public string Summary { get; set; }
+
+        [YamlMember(Alias = Constants.PropertyName.Documentation)]
+        [JsonProperty(Constants.PropertyName.Documentation)]
         [MergeOption(MergeOption.Ignore)]
-        public string HtmlId { get; set; }
+        public SourceDetail Documentation { get; set; }
 
         [ExtensibleMember]
         [JsonExtensionData]
