+++ src/Microsoft.DocAsCode.DataContracts.RestApi/RestApiResponseExampleViewModel.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.DocAsCode.Build.RestApi.ViewModels
+namespace Microsoft.DocAsCode.DataContracts.RestApi
 {
     using System;
