+++ src/Microsoft.DocAsCode.DataContracts.RestApi/RestApiResponseViewModel.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.DocAsCode.Build.RestApi.ViewModels
+namespace Microsoft.DocAsCode.DataContracts.RestApi
 {
     using System;
     using System.Collections.Generic;
@@ -10,13 +10,27 @@ namespace Microsoft.DocAsCode.Build.RestApi.ViewModels
     using YamlDotNet.Serialization;
 
     using Microsoft.DocAsCode.YamlSerialization;
+    using Microsoft.DocAsCode.Utility.EntityMergers;
 
     [Serializable]
-    public class RestApiParameterViewModel
+    public class RestApiResponseViewModel
     {
+        [YamlMember(Alias = "statusCode")]
+        [JsonProperty("statusCode")]
+        [MergeOption(MergeOption.MergeKey)]
+        public string HttpStatusCode { get; set; }
+
         [YamlMember(Alias = "description")]
         [JsonProperty("description")]
         public string Description { get; set; }
+
+        [YamlMember(Alias = "summary")]
+        [JsonProperty("summary")]
+        public string Summary { get; set; }
+
+        [YamlMember(Alias = "examples")]
+        [JsonProperty("examples")]
+        public List<RestApiResponseExampleViewModel> Examples { get; set; }
 
         [ExtensibleMember]
         [JsonExtensionData]
