+++ src/Microsoft.DocAsCode.DataContracts.RestApi/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net452" />
+  <package id="YamlDotNet.Signed" version="3.7.0" targetFramework="net452" />
+</packages>
