+++ src/Microsoft.DocAsCode.DataContracts.RestApi/project.json
@@ -1,15 +1,10 @@
 ﻿{
   "version": "1.0.0",
   "dependencies": {
-    "Microsoft.Composition": "1.0.27",
-    "Microsoft.DocAsCode.Build.Common": "1.0.0",
-    "Microsoft.DocAsCode.Common": "1.0.0",
     "Microsoft.DocAsCode.DataContracts.Common": "1.0.0",
-    "Microsoft.DocAsCode.DataContracts.RestApi": "1.0.0",
-    "Microsoft.DocAsCode.Plugins": "1.0.0",
     "Microsoft.DocAsCode.Utility": "1.0.0",
+    "Microsoft.DocAsCode.YamlSerialization": "1.0.0",
     "Newtonsoft.Json": "7.0.1",
-    "System.Collections.Immutable": "1.1.37",
     "YamlDotNet.Signed": "3.7.0"
   },
   "frameworks": {
