+++ test/Microsoft.DocAsCode.Build.RestApi.Tests/Microsoft.DocAsCode.Build.RestApi.Tests.csproj
@@ -143,6 +143,10 @@
       <Project>{60f21edb-37f9-490d-b1b6-2ec482f76b50}</Project>
       <Name>Microsoft.DocAsCode.DataContracts.Common</Name>
     </ProjectReference>
+    <ProjectReference Include="..\..\src\Microsoft.DocAsCode.DataContracts.RestApi\Microsoft.DocAsCode.DataContracts.RestApi.csproj">
+      <Project>{c6844e9d-9c68-4947-84da-a0eaa5191664}</Project>
+      <Name>Microsoft.DocAsCode.DataContracts.RestApi</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\src\Microsoft.DocAsCode.Dfm\Microsoft.DocAsCode.Dfm.csproj">
       <Project>{c212545a-62f4-40a3-a814-2ee4c7df7d41}</Project>
       <Name>Microsoft.DocAsCode.Dfm</Name>
